refactor: searchbar-separate-file
